pyproject.toml
[project]
name = "mnemosyne"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.57.0",
    "httpx>=0.28.1",
    "numpy<2.0",
    "opencv-python==4.9.0.80",
    "opencv-python-headless==4.9.0.80",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "ultralytics>=8.4.6",
    "uvicorn>=0.40.0",
]
